- zmiana srodowiska uruchomieniowego na g++ zamiast MinGw (tak bylo dziwnie jakos) - poprawienie time.h - pod linuxem stara metoda z funkcja time() czasami miala dokladnosc sekundowa zamiast mili czy nawet lepsza... - wypisywanie czasu w debugach... - drobne bledy poprawione w StegLACK - debugowanie troche poukladane - ten poziom debugow co jest teraz wydaje sie byc odpowiedni do prezentacji - dodanie operacji na plikach w Util - zeby dzialalo w koncu! - pare zmian w konfigach

+++ trunk/src/Main/Config.h
@@ -58,7 +58,6 @@ public:
 	/**
 	 * ścieżka do pliku w którym zapisujemy dane audio przychodzące.
 	 * tylko w przypadku odbierania połączenia
-	 * TODO: czy w przyadku dzwonienia tez zapisywac dane?
 	 */
 	string outputAudioFilePath;
 
@@ -101,7 +100,7 @@ public:
 	 * Czas pomiedzy odczytywaniem pojednczych pakietow z kolejki pakietow przychodzacych
 	 * (razem z rozmiarem kolejki definiuje model kolejki pakietow przychodzacych)
 	 */
-	int incQueueReadInterval;
+	long incQueueReadInterval;
 
 	/**
 	 * Rozmiar kolejki pakietów przychodzacych
